Cambio implementación de métodos en Empleado por los del SalarioFamiliar

+++ src/ar/edu/unp/po2/tp2/salarioFamiliar/SalarioFamiliar.java
@@ -2,21 +2,21 @@ package ar.edu.unp.po2.tp2.salarioFamiliar;
 
 public class SalarioFamiliar {
 	
-	private long cantidadAsignacionXHijo = 150l;
+	private double cantidadAsignacionXHijo = 150l;
 	
-	private long cantidadSiTieneConyuge = 100l;
+	private double cantidadSiTieneConyuge = 100l;
 	
-	private long montoPorAntiguedad = 50l;
+	private double montoPorAntiguedad = 50l;
 
-	public long calcularMontoAsignacionPorHijo(int cantidadHijos) {
+	public double calcularMontoAsignacionPorHijo(int cantidadHijos) {
 		return this.cantidadAsignacionXHijo * cantidadHijos;
 	}
 	
-	public long calcularMontoAsignacionPorConyuge(boolean tieneConyuge) {
+	public double calcularMontoAsignacionPorConyuge(boolean tieneConyuge) {
 		return tieneConyuge ? this.cantidadSiTieneConyuge : 0l;
 	}
 	
-	public long calcularMontoPorAntiguedad(int antiguedad) {
+	public double calcularMontoPorAntiguedad(int antiguedad) {
 		return montoPorAntiguedad * antiguedad;
 	}
 }
